update backend structure
