+++ dpnp/backend/extensions/fft/CMakeLists.txt
@@ -27,6 +27,7 @@
 set(python_module_name _fft_impl)
 set(_module_src
     ${CMAKE_CURRENT_SOURCE_DIR}/fft_py.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/c2c.cpp
 )
 
 pybind11_add_module(${python_module_name} MODULE ${_module_src})
